// common/enums.ts
export enum UserRole {
  ADMIN = "admin",
  TRAINER = "trainer",
  PLAYER = "player",
}
export enum MembershipType {
  PREMIUM = "premium",
  GOLD = "GOLD",
  BASIC = "basic",
}
export enum statusRequest {
  IDLE = "idle",
  LOADING = "loading",
  SUCCEEDED = "succeeded",
  FAILED = "failed",
}
export enum gender {
  MALE = "male",
  FEMALE = "female",
}